+++ mindspore/lite/src/ops/fill.cc
@@ -52,10 +52,9 @@ int Fill::InferShape(std::vector<tensor::Tensor *> inputs_, std::vector<tensor::
   }
 
   std::vector<int> output_shape;
-  for (int i = 0; i < GetDims().size(); i++) {
+  for (size_t i = 0; i < GetDims().size(); i++) {
     output_shape.push_back(GetDims()[i]);
   }
-//  (void)output_shape.insert(output_shape.begin(), GetDims().begin(), GetDims().end());
   output->set_shape(output_shape);
   return RET_OK;
 }
